@@ -2,14 +2,10 @@
   "name": "seamless-admin",
   "version": "1.0.0",
   "description": "A seamless django like admin panel setup for Laravel",
+  "type": "module",
   "scripts": {
-    "dev": "npm run development",
-    "development": "mix",
-    "watch": "mix watch",
-    "watch-poll": "mix watch -- --watch-options-poll=1000",
-    "hot": "mix watch --hot",
-    "prod": "npm run production",
-    "production": "mix --production"
+    "dev": "vite",
+    "build": "vite build"
   },
   "keywords": [
     "laravel",
@@ -18,14 +14,19 @@
   "author": "Advaith A J",
   "license": "MIT",
   "devDependencies": {
-    "autoprefixer": "^10.4.4",
-    "laravel-mix": "^6.0.49",
-    "postcss": "^8.4.12",
-    "resolve-url-loader": "^5.0.0",
-    "sass": "^1.67.0",
-    "sass-loader": "^12.6.0",
-    "tailwindcss": "^3.0.23",
-    "vue": "^3.2.31",
-    "vue-loader": "^16.2.0"
+    "@vitejs/plugin-vue": "^5.0.4",
+    "autoprefixer": "^10.4.18",
+    "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.0",
+    "laravel-vite-plugin": "^1.0.2",
+    "lucide-vue-next": "^0.344.0",
+    "postcss": "^8.4.35",
+    "radix-vue": "^1.4.9",
+    "sass": "^1.71.1",
+    "tailwind-merge": "^2.2.1",
+    "tailwindcss": "^3.4.1",
+    "tailwindcss-animate": "^1.0.7",
+    "vite": "^5.1.4",
+    "vue": "^3.4.21"
   }
 }
