@@ -1,1 +1,12 @@
-(()=>{var e,n={726:()=>{var e,n;window.hasOwnProperty("feather")&&window.feather.replace(),null===(e=document.getElementById("hamburger"))||void 0===e||e.addEventListener("click",(function(){var e,n;null===(e=document.getElementById("sidebar-container"))||void 0===e||null===(e=e.classList)||void 0===e||e.add("active"),null===(n=document.getElementById("sidebar-backdrop"))||void 0===n||null===(n=n.classList)||void 0===n||n.add("active")})),null===(n=document.getElementById("sidebar-backdrop"))||void 0===n||n.addEventListener("click",(function(){var e,n;null===(e=document.getElementById("sidebar-container"))||void 0===e||null===(e=e.classList)||void 0===e||e.remove("active"),null===(n=document.getElementById("sidebar-backdrop"))||void 0===n||null===(n=n.classList)||void 0===n||n.remove("active")}))},822:()=>{}},r={};function t(e){var a=r[e];if(void 0!==a)return a.exports;var i=r[e]={exports:{}};return n[e](i,i.exports,t),i.exports}t.m=n,e=[],t.O=(n,r,a,i)=>{if(!r){var o=1/0;for(v=0;v<e.length;v++){for(var[r,a,i]=e[v],d=!0,l=0;l<r.length;l++)(!1&i||o>=i)&&Object.keys(t.O).every((e=>t.O[e](r[l])))?r.splice(l--,1):(d=!1,i<o&&(o=i));if(d){e.splice(v--,1);var s=a();void 0!==s&&(n=s)}}return n}i=i||0;for(var v=e.length;v>0&&e[v-1][2]>i;v--)e[v]=e[v-1];e[v]=[r,a,i]},t.o=(e,n)=>Object.prototype.hasOwnProperty.call(e,n),(()=>{var e={773:0,170:0};t.O.j=n=>0===e[n];var n=(n,r)=>{var a,i,[o,d,l]=r,s=0;if(o.some((n=>0!==e[n]))){for(a in d)t.o(d,a)&&(t.m[a]=d[a]);if(l)var v=l(t)}for(n&&n(r);s<o.length;s++)i=o[s],t.o(e,i)&&e[i]&&e[i][0](),e[i]=0;return t.O(v)},r=self.webpackChunkseamless_admin=self.webpackChunkseamless_admin||[];r.forEach(n.bind(null,0)),r.push=n.bind(null,r.push.bind(r))})(),t.O(void 0,[170],(()=>t(726)));var a=t.O(void 0,[170],(()=>t(822)));a=t.O(a)})();
+if (window.hasOwnProperty('feather'))
+    window.feather.replace()
+
+document.getElementById('hamburger')?.addEventListener('click', () => {
+    document.getElementById('sidebar-container')?.classList?.add('active');
+    document.getElementById('sidebar-backdrop')?.classList?.add('active');
+});
+
+document.getElementById('sidebar-backdrop')?.addEventListener('click', () => {
+    document.getElementById('sidebar-container')?.classList?.remove('active');
+    document.getElementById('sidebar-backdrop')?.classList?.remove('active');
+})
