@@ -10,10 +10,10 @@
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
 <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&display=swap" rel="stylesheet" />
 {{--custom css--}}
-<link href="{{ asset('seamless-admin/css/app.css') }}" rel="stylesheet" />
+@vite('src/resources/assets/scss/app.scss', 'seamless-admin')
 
 {{--scripts--}}
 <script src="https://unpkg.com/feather-icons@4.29.0/dist/feather.min.js" defer></script>
-<script src="{{ asset('seamless-admin/js/app.js') }}" defer></script>
+@vite('src/resources/assets/js/app.js', 'seamless-admin')
 
 @yield('header')
