@@ -8,11 +8,11 @@
 @section('title', "Manage $name")
 
 @section('header')
-    <script src="{{ asset('seamless-admin/js/type-index.js') }}" defer></script>
+    @vite('src/resources/assets/js/type-index.js', 'seamless-admin')
 @endsection
 
 @section('content')
-    <div id="app" class="container mx-auto px-4 py-2">
+    <div class="container mx-auto px-4 py-2">
         <div class="flex justify-between items-center">
             <h2 class="text-2xl font-semibold">
                 Manage {{ $name }}
@@ -26,17 +26,19 @@
             @endif
         </div>
 
-        <type-index
-            data-fetch-url="{{ route('api.admin.type.type_index_data', request()->type) }}"
-            key-name="{{ $instance->getKeyName() }}"
-            v-bind:fillable="{!! str_replace('"', '\'', json_encode($instance->adminIndexFields())) !!}"
-            v-bind:can-edit="{{ $instance->adminCanAccessEdit() }}"
-            v-bind:can-delete="{{ $instance->adminCanAccessDelete() }}"
-            v-bind:routes="{
-                'show': '{{ route('admin.type.show', [request()->type, '%key%']) }}',
-                'edit': '{{ route('admin.type.edit', [request()->type, '%key%']) }}',
-                'delete': '{{ route('admin.type.delete', [request()->type]) }}'
-            }"
-        >Loading...</type-index>
+        <div id="app">
+            <type-index
+                data-fetch-url="{{ route('api.admin.type.type_index_data', request()->type) }}"
+                key-name="{{ $instance->getKeyName() }}"
+                v-bind:fillable="{!! str_replace('"', '\'', json_encode($instance->adminIndexFields())) !!}"
+                v-bind:can-edit="{{ $instance->adminCanAccessEdit() }}"
+                v-bind:can-delete="{{ $instance->adminCanAccessDelete() }}"
+                v-bind:routes="{
+                    'show': '{{ route('admin.type.show', [request()->type, '%key%']) }}',
+                    'edit': '{{ route('admin.type.edit', [request()->type, '%key%']) }}',
+                    'delete': '{{ route('admin.type.delete', [request()->type]) }}'
+                }"
+            >Loading...</type-index>
+        </div>
     </div>
 @endsection
